Refactor file structure and module names
